Stop logs loop, plus some refactor.
